+++ web/components/shared/layout/navBarV2.tsx
@@ -1,4 +1,4 @@
-import { Popover, Transition } from "@headlessui/react";
+import { Dialog, Popover, Transition } from "@headlessui/react";
 import {
   BanknotesIcon,
   Bars3Icon,
@@ -12,6 +12,7 @@ import Link from "next/link";
 import { useRouter } from "next/router";
 import Image from "next/image";
 import { Fragment, useState } from "react";
+import OnboardingButton from "../auth/onboardingButton";
 
 const navigation = [
   {
@@ -59,159 +60,104 @@ interface NavBarV2Props {
 const NavBarV2 = (props: NavBarV2Props) => {
   const { setOpenLogin, setOpenOnboarding } = props;
 
+  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
   const router = useRouter();
 
   return (
-    <header>
-      <Popover className="relative bg-white">
-        <div className="mx-auto flex max-w-7xl items-center justify-between p-6 md:justify-start md:space-x-10 lg:px-8 h-16 md:h-24">
-          <div className="flex justify-start lg:w-0 lg:flex-1">
-            <button
-              onClick={() => router.push("/")}
-              className="hidden sm:block"
-            >
-              <Image
-                className="rounded-md"
-                src="/assets/logos/logo-transparent.png"
-                width={200}
-                height={125}
-                alt="Helicone-full-logo"
-              />
-            </button>
-            <button
-              onClick={() => router.push("/")}
-              className="block sm:hidden -ml-8"
-            >
-              <Image
-                className="rounded-md"
-                src="/assets/logos/logo-blocks-transparent.png"
-                width={75}
-                height={50}
-                alt="Helicone-full-logo"
-              />
-            </button>
-          </div>
-          <div className="-my-2 -mr-2 md:hidden">
-            <Popover.Button className="inline-flex items-center justify-center rounded-md bg-white p-2 text-gray-400 hover:bg-gray-100 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
-              <span className="sr-only">Open menu</span>
-              <Bars3Icon className="h-6 w-6" aria-hidden="true" />
-            </Popover.Button>
-          </div>
-          <Popover.Group as="nav" className="hidden space-x-10 md:flex">
-            {navigation.map((item) => (
-              <Link
-                href={item.href as string}
-                target={item.target}
-                key={`desktop-${item.name}`}
-                className="text-base font-medium text-gray-500 hover:text-gray-900"
-                rel="noopener noreferrer"
-              >
-                {item.name}
-              </Link>
-            ))}
-          </Popover.Group>
-          <div className="hidden items-center justify-end md:flex md:flex-1 lg:w-0">
-            <button
-              onClick={() => setOpenLogin(true)}
-              className="whitespace-nowrap text-base font-medium text-gray-500 hover:text-gray-900"
-            >
-              Sign in
-            </button>
-            <button
-              onClick={() => setOpenOnboarding(true)}
-              className="ml-8 inline-flex items-center justify-center whitespace-nowrap rounded-md border border-transparent bg-gradient-to-r from-sky-600 to-indigo-500 bg-origin-border px-4 py-2 text-base font-medium text-white shadow-sm hover:from-sky-700 hover:to-indigo-600"
-            >
-              Sign up
-            </button>
-          </div>
+    <header className="bg-gray-50 border-b border-gray-200">
+      <nav
+        className="mx-auto flex max-w-7xl items-center justify-between gap-x-6 p-4 lg:px-8"
+        aria-label="Global"
+      >
+        <div className="flex lg:flex-1">
+          <a href="/" className="-m-1.5 p-1.5">
+            <span className="sr-only">Helicone</span>
+            <Image
+              className="rounded-md"
+              src="/assets/heli-full-logo.png"
+              width={150}
+              height={75}
+              alt="Helicone-full-logo"
+            />
+          </a>
         </div>
-
-        <Transition
-          as={Fragment}
-          enter="duration-200 ease-out"
-          enterFrom="opacity-0 scale-95"
-          enterTo="opacity-100 scale-100"
-          leave="duration-100 ease-in"
-          leaveFrom="opacity-100 scale-100"
-          leaveTo="opacity-0 scale-95"
-        >
-          <Popover.Panel
-            focus
-            className="absolute inset-x-0 top-0 z-30 origin-top-right transform p-2 transition md:hidden"
+        <div className="hidden lg:flex lg:gap-x-16">
+          {navigation.map((item) => (
+            <a
+              key={item.name}
+              href={item.href}
+              className="text-sm font-semibold leading-6 text-gray-900"
+              target={item.target}
+              rel="noopener noreferrer"
+            >
+              {item.name}
+            </a>
+          ))}
+        </div>
+        <div className="flex flex-1 items-center justify-end gap-x-4">
+          <button
+            onClick={() => setOpenLogin(true)}
+            className="rounded-md bg-white px-3.5 py-2.5 text-sm font-semibold border border-gray-300 hover:bg-sky-50 text-gray-900 shadow-sm hover:text-sky-700 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-sky-500"
           >
-            <div className="divide-y-2 divide-gray-50 rounded-lg bg-white shadow-lg ring-1 ring-black ring-opacity-5">
-              <div className="px-5 pt-5 pb-6">
-                <div className="flex items-center justify-between">
-                  <div>
-                    {/* <img
-                    className="h-8 w-auto"
-                    src="https://tailwindui.com/img/logos/mark.svg?from-color=purple&from-shade=600&to-color=indigo&to-shade=600&toShade=600"
-                    alt="Your Company"
-                  /> */}
-                  </div>
-                  <div className="-mr-2">
-                    <Popover.Button className="inline-flex items-center justify-center rounded-md bg-white p-2 text-gray-400 hover:bg-gray-100 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
-                      <span className="sr-only">Close menu</span>
-                      <XMarkIcon className="h-6 w-6" aria-hidden="true" />
-                    </Popover.Button>
-                  </div>
-                </div>
-                <div className="mt-6">
-                  <nav className="grid grid-cols-1 gap-7">
-                    {navigation.map((item) => (
-                      <a
-                        key={`mobile-${item.name}`}
-                        href={item.href}
-                        className="-m-3 flex items-center rounded-lg p-3 hover:bg-gray-50"
-                      >
-                        <div className="flex h-10 w-10 flex-shrink-0 items-center justify-center rounded-md bg-gradient-to-r from-sky-600 to-indigo-500 text-white">
-                          <item.icon className="h-6 w-6" aria-hidden="true" />
-                        </div>
-                        <div className="ml-4 text-base font-medium text-gray-900">
-                          {item.name}
-                        </div>
-                      </a>
-                    ))}
-                  </nav>
-                </div>
-              </div>
-              <div className="py-6 px-5">
-                <div className="">
-                  <Popover.Panel>
-                    {({ close }) => (
-                      <button
-                        className="flex w-full items-center justify-center rounded-md border border-transparent bg-gradient-to-r from-sky-600 to-indigo-500 bg-origin-border px-4 py-2 text-base font-medium text-white shadow-sm hover:from-purple-700 hover:to-indigo-700"
-                        onClick={() => {
-                          close();
-                          setOpenOnboarding(true);
-                        }}
-                      >
-                        Sign Up
-                      </button>
-                    )}
-                  </Popover.Panel>
-                  <p className="mt-6 text-center text-base font-medium text-gray-500">
-                    Existing customer?{` `}
-                    <Popover.Panel className="inline">
-                      {({ close }) => (
-                        <button
-                          className="text-gray-900"
-                          onClick={() => {
-                            close();
-                            setOpenLogin(true);
-                          }}
-                        >
-                          Sign In
-                        </button>
-                      )}
-                    </Popover.Panel>
-                  </p>
-                </div>
+            Log In
+          </button>
+          <OnboardingButton variant="secondary" title={"Sign Up"} />
+        </div>
+        <div className="flex lg:hidden">
+          <button
+            type="button"
+            className="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-gray-700"
+            onClick={() => setMobileMenuOpen(true)}
+          >
+            <span className="sr-only">Open main menu</span>
+            <Bars3Icon className="h-6 w-6" aria-hidden="true" />
+          </button>
+        </div>
+      </nav>
+      <Dialog
+        as="div"
+        className="lg:hidden"
+        open={mobileMenuOpen}
+        onClose={setMobileMenuOpen}
+      >
+        <div className="fixed inset-0 z-10" />
+        <Dialog.Panel className="fixed inset-y-0 right-0 z-10 w-full overflow-y-auto bg-gray-50 px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10">
+          <div className="flex items-center gap-x-6 justify-between">
+            <a href="#" className="-m-1.5 p-1.5">
+              <span className="sr-only">Your Company</span>
+              <img
+                className="h-8 w-auto"
+                src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600"
+                alt=""
+              />
+            </a>
+
+            <button
+              type="button"
+              className="-m-2.5 rounded-md p-2.5 text-gray-700"
+              onClick={() => setMobileMenuOpen(false)}
+            >
+              <span className="sr-only">Close menu</span>
+              <XMarkIcon className="h-6 w-6" aria-hidden="true" />
+            </button>
+          </div>
+          <div className="mt-6 flow-root">
+            <div className="-my-6 divide-y divide-gray-500/10">
+              <div className="space-y-2 py-6">
+                {navigation.map((item) => (
+                  <a
+                    key={item.name}
+                    href={item.href}
+                    className="-mx-3 block rounded-lg py-2 px-3 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50"
+                  >
+                    {item.name}
+                  </a>
+                ))}
               </div>
             </div>
-          </Popover.Panel>
-        </Transition>
-      </Popover>
+          </div>
+        </Dialog.Panel>
+      </Dialog>
     </header>
   );
 };
