+++ web/components/templates/home/detailsV2.tsx
@@ -42,9 +42,6 @@ const Details = () => {
         <div className="grid grid-cols-1 gap-y-16 gap-x-8 sm:gap-y-20 lg:grid-cols-2 lg:items-start">
           <div className="px-6 lg:px-0 lg:pt-4 lg:pr-4">
             <div className="mx-auto max-w-2xl lg:mx-0 lg:max-w-lg">
-              <h2 className="text-lg font-semibold leading-8 tracking-tight text-sky-600">
-                One Line of Code
-              </h2>
               <p className="mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
                 Let us handle the analytics.
               </p>
