--- web/pages/test.tsx
@@ -1,367 +1,0 @@
-/* eslint-disable @next/next/no-img-element */
-import { Dialog } from "@headlessui/react";
-import {
-  Bars3Icon,
-  ChevronRightIcon,
-  XMarkIcon,
-} from "@heroicons/react/20/solid";
-import { useState } from "react";
-import Details from "../components/templates/home/detailsV2";
-
-const navigation = [
-  { name: "Pricing", href: "#" },
-  { name: "Docs", href: "#" },
-  { name: "Discord", href: "#" },
-  { name: "Github", href: "#" },
-  { name: "Contact", href: "#" },
-];
-
-const timeline = [
-  {
-    name: "Founded company",
-    description:
-      "Drawn by the entrepreneurial spirit, we quit our software engineering jobs looking to build within the generative AI space.",
-    date: "Jan 2023",
-    dateTime: "2023-01",
-  },
-  {
-    name: "Backed by YCombinator",
-    description:
-      "We were accepted into YCombinator's Winter 2023 batch and received $500k in funding.",
-    date: "Jan 2023",
-    dateTime: "2023-01",
-  },
-  {
-    name: "Launched Helicone",
-    description:
-      "After building multiple LLM apps and struggling with the lack of good monitoring and observability tools, we decided to build Helicone.",
-    date: "Feb 2022",
-    dateTime: "2023-02",
-  },
-  {
-    name: "3 million requests",
-    description:
-      "We reached 3 million requests in a single month, and we're just getting started.",
-    date: "Mar 2023",
-    dateTime: "2023-03",
-  },
-];
-
-export default function Example() {
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-  return (
-    <>
-      <div className="relative isolate overflow-hidden bg-gray-50">
-        <svg
-          className="absolute inset-0 -z-10 h-full w-full stroke-gray-200 [mask-image:radial-gradient(100%_100%_at_top_right,white,transparent)]"
-          aria-hidden="true"
-        >
-          <defs>
-            <pattern
-              id="0787a7c5-978c-4f66-83c7-11c213f99cb7"
-              width={200}
-              height={200}
-              x="50%"
-              y={-1}
-              patternUnits="userSpaceOnUse"
-            >
-              <path d="M.5 200V.5H200" fill="none" />
-            </pattern>
-          </defs>
-          <rect
-            width="100%"
-            height="100%"
-            strokeWidth={0}
-            fill="url(#0787a7c5-978c-4f66-83c7-11c213f99cb7)"
-          />
-        </svg>
-        <header className="bg-gray-50 border-b border-gray-200">
-          <nav
-            className="mx-auto flex max-w-7xl items-center justify-between gap-x-6 p-4 lg:px-8"
-            aria-label="Global"
-          >
-            <div className="flex lg:flex-1">
-              <a href="#" className="-m-1.5 p-1.5">
-                <span className="sr-only">Your Company</span>
-                <img
-                  className="h-8 w-auto"
-                  src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600"
-                  alt=""
-                />
-              </a>
-            </div>
-            <div className="hidden lg:flex lg:gap-x-16">
-              {navigation.map((item) => (
-                <a
-                  key={item.name}
-                  href={item.href}
-                  className="text-sm font-semibold leading-6 text-gray-900"
-                >
-                  {item.name}
-                </a>
-              ))}
-            </div>
-            <div className="flex flex-1 items-center justify-end gap-x-6">
-              <a
-                href="#"
-                className="hidden lg:block lg:text-sm lg:font-semibold lg:leading-6 lg:text-gray-900"
-              >
-                Log in
-              </a>
-              <a
-                href="#"
-                className="rounded-md bg-sky-600 py-2 px-3 text-sm font-semibold text-white shadow-sm hover:bg-sky-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-sky-600"
-              >
-                Sign up
-              </a>
-            </div>
-            <div className="flex lg:hidden">
-              <button
-                type="button"
-                className="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-gray-700"
-                onClick={() => setMobileMenuOpen(true)}
-              >
-                <span className="sr-only">Open main menu</span>
-                <Bars3Icon className="h-6 w-6" aria-hidden="true" />
-              </button>
-            </div>
-          </nav>
-          <Dialog
-            as="div"
-            className="lg:hidden"
-            open={mobileMenuOpen}
-            onClose={setMobileMenuOpen}
-          >
-            <div className="fixed inset-0 z-10" />
-            <Dialog.Panel className="fixed inset-y-0 right-0 z-10 w-full overflow-y-auto bg-gray-50 px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10">
-              <div className="flex items-center gap-x-6">
-                <a href="#" className="-m-1.5 p-1.5">
-                  <span className="sr-only">Your Company</span>
-                  <img
-                    className="h-8 w-auto"
-                    src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600"
-                    alt=""
-                  />
-                </a>
-                <a
-                  href="#"
-                  className="ml-auto rounded-md bg-indigo-600 py-2 px-3 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-                >
-                  Sign up
-                </a>
-                <button
-                  type="button"
-                  className="-m-2.5 rounded-md p-2.5 text-gray-700"
-                  onClick={() => setMobileMenuOpen(false)}
-                >
-                  <span className="sr-only">Close menu</span>
-                  <XMarkIcon className="h-6 w-6" aria-hidden="true" />
-                </button>
-              </div>
-              <div className="mt-6 flow-root">
-                <div className="-my-6 divide-y divide-gray-500/10">
-                  <div className="space-y-2 py-6">
-                    {navigation.map((item) => (
-                      <a
-                        key={item.name}
-                        href={item.href}
-                        className="-mx-3 block rounded-lg py-2 px-3 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50"
-                      >
-                        {item.name}
-                      </a>
-                    ))}
-                  </div>
-                  <div className="py-6">
-                    <a
-                      href="#"
-                      className="-mx-3 block rounded-lg py-2.5 px-3 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50"
-                    >
-                      Log in
-                    </a>
-                  </div>
-                </div>
-              </div>
-            </Dialog.Panel>
-          </Dialog>
-        </header>
-        <div className="mx-auto max-w-7xl px-6 pt-10 pb-24 sm:pb-32 lg:flex lg:py-28 lg:px-8">
-          <div className="mx-auto max-w-2xl lg:mx-0 lg:max-w-xl lg:flex-shrink-0 lg:pt-8">
-            {/* <img
-            className="h-11"
-            src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600"
-            alt="Your Company"
-          /> */}
-            <div className="mt-24 sm:mt-32 lg:mt-8">
-              <a href="#" className="inline-flex space-x-6">
-                <span className="rounded-full bg-orange-600/10 px-3 py-1 text-sm font-semibold leading-6 text-orange-600 ring-1 ring-inset ring-orange-600/10">
-                  Backed by YCombinator
-                </span>
-                <span className="sm:inline-flex hidden items-center space-x-2 text-sm font-medium leading-6 text-gray-600">
-                  <span>We&apos;re Open-Source</span>
-                  <ChevronRightIcon
-                    className="h-5 w-5 text-gray-400"
-                    aria-hidden="true"
-                  />
-                </span>
-              </a>
-            </div>
-            <h1 className="mt-8 text-4xl font-bold tracking-tight text-gray-900 sm:text-7xl">
-              Observability for LLM applications
-            </h1>
-            <p className="mt-6 text-xl leading-8 text-gray-600">
-              Track usage, costs, and latency metrics with one line of code.
-              We&apos;re an open-source observability platform that helps you
-              better understand your GPT-3 application.
-            </p>
-            <div className="mt-10 flex items-center gap-x-6">
-              <a
-                href="#"
-                className="rounded-md bg-sky-500 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-sky-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-sky-500"
-              >
-                Get started
-              </a>
-              <a
-                href="#"
-                className="text-sm font-semibold leading-6 text-gray-900"
-              >
-                Learn more <span aria-hidden="true">→</span>
-              </a>
-            </div>
-          </div>
-          <div className="mx-auto mt-16 flex max-w-2xl sm:mt-24 lg:ml-32 lg:mt-0 lg:mr-0 lg:max-w-none lg:flex-none xl:ml-32">
-            <div className="max-w-3xl flex-none sm:max-w-5xl lg:max-w-none">
-              <div className="-m-2 rounded-xl bg-gray-900/5 p-2 ring-1 ring-inset ring-gray-900/10 lg:-m-4 lg:rounded-2xl lg:p-4">
-                <img
-                  src="/assets/test.png"
-                  alt="App screenshot"
-                  width={2432}
-                  height={1442}
-                  className="w-[76rem] rounded-md shadow-2xl ring-1 ring-gray-900/10"
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      {/* FEATURE */}
-      <div className="bg-white py-24 sm:py-16">
-        <div className="mx-auto max-w-7xl px-6 lg:px-8">
-          <div className="mx-auto max-w-2xl lg:mx-0">
-            <h2 className="text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
-              Even wonder how much you spend on GPT-3? We do too.
-            </h2>
-            <p className="mt-6 text-base leading-7 text-gray-600">
-              Creating an internal tool for observability and monitoring is
-              difficult to build and even harder to maintain. And general
-              logging tools like Mixpanel or Datadog aren&apos;t tailored enough
-              for LLMs. Helicone is a platform built specifically for developers
-              using LLMs.
-            </p>
-          </div>
-          <div className="mx-auto mt-16 flex max-w-2xl flex-col gap-8 lg:mx-0 lg:mt-20 lg:max-w-none lg:flex-row lg:items-end">
-            <div className="flex flex-col-reverse justify-between gap-x-16 gap-y-8 rounded-2xl bg-gray-100 p-8 sm:w-3/4 sm:max-w-md sm:flex-row-reverse sm:items-end lg:w-72 lg:max-w-none lg:flex-none lg:flex-col lg:items-start">
-              <p className="flex-none text-3xl font-bold tracking-tight text-gray-900">
-                50+
-              </p>
-              <div className="sm:w-80 sm:shrink lg:w-auto lg:flex-none">
-                <p className="text-lg font-semibold tracking-tight text-gray-900">
-                  Users on the platform
-                </p>
-                <p className="mt-2 text-base leading-7 text-gray-600">
-                  We provide cloud and self-hosting solutions
-                </p>
-              </div>
-            </div>
-            <div className="flex flex-col-reverse justify-between gap-y-8 gap-x-16 rounded-2xl bg-gray-900 p-8 sm:flex-row-reverse sm:items-end lg:w-full lg:max-w-sm lg:flex-auto lg:flex-col lg:items-start lg:gap-y-44">
-              <p className="flex-none text-3xl font-bold tracking-tight text-white">
-                3.7 Million
-              </p>
-              <div className="sm:w-80 sm:shrink lg:w-auto lg:flex-none">
-                <p className="text-lg font-semibold tracking-tight text-white">
-                  We&apos;re proud that our customers have made over 3 million
-                  requests.
-                </p>
-                <p className="mt-2 text-base leading-7 text-gray-400">
-                  We currently support GPT-3, ChatGPT, GPT-4, and OpenAI&apos;s
-                  moderations.
-                </p>
-              </div>
-            </div>
-            <div className="flex flex-col-reverse justify-between gap-y-8 gap-x-16 rounded-2xl bg-sky-600 p-8 sm:w-11/12 sm:max-w-xl sm:flex-row-reverse sm:items-end lg:w-full lg:max-w-none lg:flex-auto lg:flex-col lg:items-start lg:gap-y-28">
-              <p className="flex-none text-3xl font-bold tracking-tight text-white">
-                &gt;350,000
-              </p>
-              <div className="sm:w-80 sm:shrink lg:w-auto lg:flex-none">
-                <p className="text-lg font-semibold tracking-tight text-white">
-                  Requests per day
-                </p>
-                <p className="mt-2 text-base leading-7 text-indigo-200">
-                  Every day our customers are getting value from Helicone.
-                </p>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      {/* DetailsV2 */}
-      <Details />
-      {/* Open Source */}
-      <div className="bg-white py-24 sm:py-32">
-        <div className="flex flex-col sm:flex-row w-full justify-between mx-auto max-w-7xl px-6 lg:px-8 items-start sm:items-center space-y-6 sm:space-y-0">
-          <div className="mx-auto max-w-2xl lg:mx-0">
-            <h2 className="text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">
-              Join our fast-growing open source community
-            </h2>
-            <p className="mt-6 text-base leading-7 text-gray-600">
-              We&apos;ve been developer driven from the start, and we&apos;re
-              committed to keeping it that way. We&apos;re always looking for
-              new contributors to help us build the best open source developer
-              tools.
-            </p>
-          </div>
-          <div className="block space-x-4">
-            <button className="rounded-md bg-white px-3.5 py-2.5 text-sm font-semibold border border-gray-300 hover:bg-sky-50 text-gray-900 shadow-sm hover:text-sky-700 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-sky-500">
-              Join Discord
-            </button>
-            <button className="rounded-md bg-sky-500 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-sky-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-sky-500">
-              View Github
-            </button>
-          </div>
-        </div>
-        <div className="py-8 sm:py-16">
-          <div className="mx-auto max-w-7xl px-6 lg:px-8">
-            <div className="mx-auto grid max-w-2xl grid-cols-1 gap-8 overflow-hidden lg:mx-0 lg:max-w-none lg:grid-cols-4">
-              {timeline.map((item) => (
-                <div key={item.name}>
-                  <time
-                    dateTime={item.dateTime}
-                    className="flex items-center text-sm font-semibold leading-6 text-sky-600"
-                  >
-                    <svg
-                      viewBox="0 0 4 4"
-                      className="mr-4 h-1 w-1 flex-none"
-                      aria-hidden="true"
-                    >
-                      <circle cx={2} cy={2} r={2} fill="currentColor" />
-                    </svg>
-                    {item.date}
-                    <div
-                      className="absolute -ml-2 h-px w-screen -translate-x-full bg-gray-900/10 sm:-ml-4 lg:static lg:ml-8 lg:-mr-6 lg:w-auto lg:flex-auto lg:translate-x-0"
-                      aria-hidden="true"
-                    />
-                  </time>
-                  <p className="mt-6 text-lg font-semibold leading-8 tracking-tight text-gray-900">
-                    {item.name}
-                  </p>
-                  <p className="mt-1 text-base leading-7 text-gray-600">
-                    {item.description}
-                  </p>
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
-      </div>
-    </>
-  );
-}
